Rename download.rs to backend.rs
This will enable to split out the contents of lib.rs into different
modules.
